refactor: RacingRequest 패키지 이동 서비스 레이어에서 필요해보인다.
